Core/Stats: Fix all player base stats - Proper table player_levelstats with blizz informations - Update saved stats to a uint16 instead of uint8

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -3375,7 +3375,7 @@ void ObjectMgr::LoadPlayerInfo()
 
             for (int i = 0; i < MAX_STATS; i++)
             {
-                pLevelInfo->stats[i] = fields[i+3].GetUInt8();
+                pLevelInfo->stats[i] = fields[i+3].GetUInt16();
             }
 
             ++count;
